Changed the file structure for signup screens
