+++ app/src/main/java/com/it235/nureserved/MainActivity.kt
@@ -24,9 +24,9 @@ import com.it235.nureserved.ui.screens.RoomUsageRules
 import com.it235.nureserved.ui.screens.SplashScreen
 import com.it235.nureserved.ui.screens.TermsAndConditionsScreen
 import com.it235.nureserved.ui.screens.authscreenui.LoginScreen
-import com.it235.nureserved.ui.screens.authscreenui.NameSignUpScreen
-import com.it235.nureserved.ui.screens.authscreenui.ProgramStudentNumberSignUpScreen
-import com.it235.nureserved.ui.screens.authscreenui.SignUpScreen
+import com.it235.nureserved.ui.screens.authscreenui.signup.NameSignUpScreen
+import com.it235.nureserved.ui.screens.authscreenui.signup.ProgramStudentNumberSignUpScreen
+import com.it235.nureserved.ui.screens.authscreenui.signup.SignUpScreen
 import com.it235.nureserved.ui.screens.homesreenui.HomeScreen
 import com.it235.nureserved.ui.screens.homesreenui.RoomDetails
 import com.it235.nureserved.ui.theme.NUreservedTheme
